+++ core/components/filedownloadr/src/Snippets/FileDownloadLinkSnippet.php
@@ -28,14 +28,14 @@ class FileDownloadLinkSnippet extends Snippet
             'ajaxMode::bool' => false,
             'chkDesc' => '',
             'countDownloads::bool' => true,
+            'countUserDownloads::bool' => false,
             'dateFormat' => 'Y-m-d',
             'directLink::bool' => false,
-            'fdlid' => '',
-            'fileCss' => $this->filedownloadr->getOption('assets_url') . 'css/fd.min.css',
+            'fileCss' => '{fd_assets_url}css/fd.min.css',
             'fileJs' => '',
             'geoApiKey' => $this->filedownloadr->getOption('ipinfodb_api_key'),
             'getFile' => '',
-            'imgLocat' => $this->filedownloadr->getOption('assets_url') . 'img/filetypes/',
+            'imgLocat' => '{fd_assets_url}img/filetypes/',
             'imgTypes' => 'fdImages',
             'mediaSourceId::int' => 0,
             'noDownload::bool' => false,
@@ -56,52 +56,37 @@ class FileDownloadLinkSnippet extends Snippet
      */
     public function execute()
     {
-        $this->filedownloadr->setConfigs($this->getProperties());
+        $this->filedownloadr->initConfig($this->getProperties());
 
         if (!$this->filedownloadr->isAllowed()) {
             return $this->filedownloadr->parse->getChunk($this->getProperty('tplNotAllowed'), $this->getProperties());
         }
         if ($this->getProperty('fileCss') !== 'disabled') {
-            $this->modx->regClientCSS($this->filedownloadr->replacePropPhs($this->getProperty('fileCss')));
+            $this->modx->regClientCSS($this->getProperty('fileCss'));
         }
         if ($this->getProperty('ajaxMode') && !empty($this->getProperty('ajaxControllerPage'))) {
             if ($this->getProperty('fileJs') !== 'disabled') {
-                $this->modx->regClientStartupScript($this->filedownloadr->replacePropPhs($this->getProperty('fileJs')));
+                $this->modx->regClientStartupScript($this->getProperty('fileJs'));
             }
         }
 
+        // Check the referrer if a file is downloaded
         if (!empty($_GET['fdlfile'])) {
-            $ref = $_SERVER['HTTP_REFERER'];
-            // deal with multiple snippets which have &browseDirectories
-            $xRef = @explode('?', $ref);
-            $queries = [];
-            parse_str($xRef[1], $queries);
-            if (!empty($queries['id'])) {
-                // non FURL
-                $baseRef = $xRef[0] . '?id=' . $queries['id'];
-            } else {
-                $baseRef = $xRef[0];
-            }
-            $baseRef = urldecode($baseRef);
-            $page = $this->modx->makeUrl($this->modx->resource->get('id'), '', '', 'full');
-            // check referrer and the page
-            if ($baseRef !== $page) {
-                $this->modx->sendUnauthorizedPage();
-            }
+            $this->filedownloadr->checkReferrer();
         }
 
         if (!$this->getProperty('downloadByOther')) {
             $sanitizedGets = $this->modx->sanitize($_GET);
             if (!empty($sanitizedGets['fdlfile'])) {
-                if (!$this->filedownloadr->checkHash($this->modx->context->key, $sanitizedGets['fdlfile'])) {
-                    return '';
+                // Download file
+                $file = $sanitizedGets['fdlfile'];
+                if ($this->filedownloadr->checkHash($this->modx->context->key, $file)) {
+                    $this->filedownloadr->downloadFile($file);
+                    // Simply terminate, because this is a downloading state
+                    @session_write_close();
+                    exit();
                 }
-                if (!$this->filedownloadr->downloadFile($sanitizedGets['fdlfile'])) {
-                    return '';
-                }
-                // simply terminate, because this is a downloading state
-                @session_write_close();
-                exit();
+                return '';
             }
         }
 
@@ -110,24 +95,15 @@ class FileDownloadLinkSnippet extends Snippet
             return '';
         }
 
+        $phs = $this->filedownloadr->config;
         $fileInfos = $contents['file'][0];
-        $tmp = [];
         foreach ($fileInfos as $k => $v) {
-            $tmp[$this->getProperty('prefix') . $k] = $v;
+            $phs[$this->getProperty('prefix') . $k] = $v;
         }
-        // fallback without prefix
-        $fileInfos = array_merge($fileInfos, $tmp);
 
-        /**
-         * for Output Filter Modifier
-         * @link http://rtfm.modx.com/display/revolution20/Custom+Output+Filter+Examples#CustomOutputFilterExamples-CreatingaCustomOutputModifier
-         */
         if (!empty($this->getProperty('input'))) {
-            $output = $fileInfos[$this->getProperty('options')];
-            // avoid 0 (zero) of the download counting.
-            if (empty($output) && !is_numeric($output)) {
-                $output = $this->filedownloadr->parse->getChunk($this->getProperty('tpl'), $fileInfos);
-            }
+            // Run as for Output Filter
+            $output = !empty($fileInfos[$this->getProperty('options')]) ? $fileInfos[$this->getProperty('options')] : '';
         } elseif ($this->getProperty('toArray')) {
             $output = '<pre>' . print_r($fileInfos, true) . '</pre>';
         } else {